requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
email-validator==2.1.0
Pillow==10.1.0
tensorflow==2.15.0
numpy==1.26.1
pandas==2.1.2
scikit-learn==1.3.2
opencv-python==4.8.1.78
geopy==2.4.1
folium==0.15.0
python-dotenv==1.0.0
gunicorn==21.2.0
google-generativeai==0.6.0
Flask-Mail==0.9.1
twilio==9.2.3
Flask-SocketIO==5.3.6
python-socketio==5.10.0